fix(mapper): only refresh cookies if online + attempt refresh on connectivity restore
