+++ src/mapper/src/routes/+layout.svelte
@@ -4,12 +4,15 @@
 
 	import { onMount } from 'svelte';
 	import { online } from 'svelte/reactivity/window';
+	import { error } from '@sveltejs/kit';
 	import type { PageProps } from './$types';
 	import { pwaInfo } from 'virtual:pwa-info';
 	import { useRegisterSW } from 'virtual:pwa-register/svelte';
 	import type { RegisterSWOptions } from 'vite-plugin-pwa/types';
 
 	import { getCommonStore, getAlertStore } from '$store/common.svelte.ts';
+	import { getLoginStore } from '$store/login.svelte.ts';
+	import { refreshCookies, getUserDetailsFromApi } from '$lib/api/login';
 	import Toast from '$lib/components/toast.svelte';
 	import Header from '$lib/components/header.svelte';
 	import OfflineBanner from '$lib/components/offline/offline-banner.svelte';
@@ -18,9 +21,13 @@
 	let { data, children }: PageProps = $props();
 
 	const commonStore = getCommonStore();
+	const loginStore = getLoginStore();
 	const alertStore = getAlertStore();
 	commonStore.setConfig(data.config);
+
 	let dbPromise = data.dbPromise;
+	let lastOnlineStatus: boolean | null = $state(null);
+	let loginDebounce: ReturnType<typeof setTimeout> | null = $state(null);
 
 	// Required for PWA to work with svelte
 	const webManifestLink = $derived(pwaInfo ? pwaInfo.webManifest.linkTag : '');
@@ -36,6 +43,55 @@
 			alertStore.setAlert({ message: m['offline.ready_offline'](), variant: 'default', duration: 2000 });
         },
     });
+
+	async function refreshCookiesAndLogin() {
+		try {
+			/*
+				Login + user details
+			*/
+			if (online.current) {
+				// Online: always go through API and refresh cookies
+				let apiUser = await refreshCookies(fetch);
+				loginStore.setRefreshCookieResponse(apiUser);
+
+				// svcfmtm is the default 'temp' user, to still allow mapping without login
+				if (apiUser?.username !== 'svcfmtm') {
+					// Call /auth/me to populate the user details in the header
+					apiUser = await getUserDetailsFromApi(fetch);
+
+					if (!apiUser) {
+						loginStore.signOut();
+						throw error(401, { message: `You must log in first` });
+					} else {
+						loginStore.setAuthDetails(apiUser);
+					}
+				}
+			}
+		} catch (error) {
+			console.warn('Error getting user login details')
+		}
+	}
+
+	// Attempt cookie refresh / login once connectivity restored
+	$effect(() => {
+		const isOnline = online.current;
+
+		// Prevent running unnecessarily
+		if (isOnline === lastOnlineStatus) return;
+		lastOnlineStatus = isOnline;
+
+		if (loginDebounce) {
+			clearTimeout(loginDebounce);
+			loginDebounce = null;
+		}
+
+		loginDebounce = setTimeout(() => {
+			if (isOnline) {
+				console.log('logging in')
+				refreshCookiesAndLogin();
+			}
+		}, 200);
+	});
 
 	onMount(async () => {
 		// Dynamically inject CSS specified in config
