+++ src/mapper/src/routes/+layout.ts
@@ -1,10 +1,7 @@
 import 'virtual:uno.css';
 import type { LayoutLoad } from './$types';
-import { error } from '@sveltejs/kit';
 
 import { getDbOnce } from '$lib/db/pglite';
-import { getLoginStore } from '$store/login.svelte.ts';
-import { refreshCookies, getUserDetailsFromApi } from '$lib/api/login';
 
 // NOTE we can't prerender as we are using dynamic routing [projectId]
 export const prerender = false;
@@ -12,26 +9,7 @@ export const ssr = false;
 
 export const load: LayoutLoad = async ({ fetch }) => {
 	let config;
-	const loginStore = getLoginStore();
 	const dbPromise = getDbOnce(); // Don't await here to allow loading spinner in layout.svelte
-
-	try {
-		/*
-			Login + user details
-		*/
-		let apiUser = await refreshCookies(fetch);
-		loginStore.setRefreshCookieResponse(apiUser);
-		if (apiUser?.username !== 'svcfmtm') {
-			// Call /auth/me to populate the user details in the header
-			apiUser = await getUserDetailsFromApi(fetch);
-			if (!apiUser) {
-				loginStore.signOut();
-				throw error(401, { message: `You must log in first` });
-			} else {
-				loginStore.setAuthDetails(apiUser);
-			}
-		}
-	} catch (error) {}
 
 	try {
 		const s3Response = await fetch(`${import.meta.env.VITE_S3_URL}/fmtm-data/frontend/config.json`);
