+++ src/mapper/src/store/login.svelte.ts
@@ -1,3 +1,5 @@
+import { refreshCookies } from '$lib/api/login';
+
 type authDetailsType = {
 	id: number;
 	username: string;
@@ -10,8 +12,6 @@ type authDetailsType = {
 	// project_roles: string | null;
 	// orgs_managed: number[];
 };
-
-import { refreshCookies } from '$lib/api/login';
 
 let authDetails: authDetailsType | null = $state(null);
 let isLoginModalOpen: boolean = $state(false);
